+++ include/nfx/Gl.h
@@ -7,5 +7,6 @@
  * Include this single header to access the full nfx::graphics::gl public API.
  */
 
+#include "nfx/graphics/gl/core/buffers/Buffer.h"
 #include "nfx/graphics/gl/core/Context.h"
 #include "nfx/graphics/gl/core/GlDefinitions.h"